refactor(orders): co-locate all non-entrypoint files under components
Moves add-detail, delivery, details, payment-history and the payments modals
(with their tests) into orders/components so the module root holds only
entrypoints and pure helpers. Refs #131.
