+++ resources/js/pages/orders/components/details.tsx
